+++ slash/log.py
@@ -60,7 +60,17 @@ class SessionLogging(object):
     def get_test_logging_context(self):
         with self._get_file_logging_context(config.root.log.subpath, config.root.log.last_test_symlink) as path:
             self.test_log_path = path
-            yield path
+            context.result.set_log_path(path)
+            try:
+                yield path
+            finally:
+                self._create_last_failed_symlink_if_needed()
+
+    def _create_last_failed_symlink_if_needed(self):
+        result = context.result
+        assert result
+        if result.is_error() or result.is_failure():
+            self._try_create_symlink(result.get_log_path(), config.root.log.last_failed_symlink)
 
     @contextmanager
     def get_session_logging_context(self):